add types for components props
