+++ packages/core/package.json
@@ -84,7 +84,8 @@
     "lodash": "^4.17.15",
     "react-is": "^16.13.1",
     "react-reconciler": "^0.25.1",
-    "tslib": "^2.0.3"
+    "tslib": "^2.0.3",
+    "utility-types": "^3.10.0"
   },
   "peerDependencies": {
     "react": ">=16.9.0"
